style (list-book): moving package and renaming class
