+++ src/main/java/br/com/zup/casadocodigo/book/BookController.java
@@ -1,8 +1,11 @@
 package br.com.zup.casadocodigo.book;
 
 import br.com.zup.casadocodigo.author.Author;
+import br.com.zup.casadocodigo.book.create.CreateBookRequest;
+import br.com.zup.casadocodigo.book.list.ListBookResponse;
 import br.com.zup.casadocodigo.category.Category;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -11,15 +14,16 @@ import org.springframework.web.bind.annotation.RestController;
 import javax.persistence.EntityManager;
 import javax.transaction.Transactional;
 import javax.validation.Valid;
+import java.util.List;
 import java.util.function.LongFunction;
 
 @RestController
 @RequestMapping("/book")
-public class CreateBookController {
+public class BookController {
 
     private final EntityManager entityManager;
 
-    public CreateBookController(EntityManager entityManager) {
+    public BookController(EntityManager entityManager) {
         this.entityManager = entityManager;
     }
